refactor: move to new class
