+++ lib/widgets/audio_bottom_sheet/audio_minimized_info.dart
@@ -1,11 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:qurantafsir_flutter/shared/constants/button_audio_enum.dart';
 import 'package:qurantafsir_flutter/shared/constants/qp_colors.dart';
 import 'package:qurantafsir_flutter/shared/constants/qp_text_style.dart';
-import 'package:qurantafsir_flutter/shared/core/providers/audio_provider.dart';
 import 'package:qurantafsir_flutter/widgets/audio_bottom_sheet/audio_bottom_sheet_state_notifier.dart';
+import 'package:qurantafsir_flutter/widgets/audio_bottom_sheet/audio_minimized_info_icon_button.dart';
 
 class AudioMinimizedInfo extends ConsumerStatefulWidget {
   const AudioMinimizedInfo({
@@ -61,7 +59,7 @@ class _AudioMinimizedInfoState extends ConsumerState<AudioMinimizedInfo> {
           children: <Widget>[
             const Padding(
               padding: EdgeInsets.all(16),
-              child: _DynamicIconButton(),
+              child: AudioMinimizedInfoIconButton(),
             ),
             Column(
               mainAxisAlignment: MainAxisAlignment.center,
@@ -87,74 +85,6 @@ class _AudioMinimizedInfoState extends ConsumerState<AudioMinimizedInfo> {
               onPressed: widget.onClose,
             ),
           ],
-        ),
-      ),
-    );
-  }
-}
-
-class _DynamicIconButton extends ConsumerWidget {
-  const _DynamicIconButton({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final AsyncValue<ButtonAudioState> buttonState =
-        ref.watch(buttonAudioStateProvider);
-
-    return buttonState.when(
-      data: (data) {
-        return InkWell(
-          onTap: () {
-            if (buttonState.value == ButtonAudioState.paused) {
-              ref.read(audioBottomSheetProvider.notifier).playAudio();
-
-              return;
-            }
-            ref.read(audioBottomSheetProvider.notifier).pauseAudio();
-          },
-          child: _IconButton(
-            icon: buttonState.value == ButtonAudioState.paused
-                ? Icons.play_arrow
-                : Icons.pause,
-          ),
-        );
-      },
-      error: (_, __) {
-        return const SizedBox.shrink();
-      },
-      loading: () {
-        return _IconButton(
-          icon: buttonState.value == ButtonAudioState.paused
-              ? Icons.play_arrow
-              : Icons.pause,
-        );
-      },
-    );
-  }
-}
-
-class _IconButton extends StatelessWidget {
-  const _IconButton({
-    Key? key,
-    required this.icon,
-  }) : super(key: key);
-
-  final IconData icon;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 36,
-      width: 36,
-      decoration: const BoxDecoration(
-        color: QPColors.brandFair,
-        shape: BoxShape.circle,
-      ),
-      child: Center(
-        child: Icon(
-          icon,
-          color: Colors.white,
-          size: 20,
         ),
       ),
     );
